+++ local_test_settings.py
@@ -1,5 +1,11 @@
 from .base import *
 
+
 INSTALLED_APPS += (
+    'wagtail_personalisation',
+    'wagtailfontawesome',
     'wagtailsurveys',
 )
+
+CELERY_ALWAYS_EAGER = True
+BROKER_BACKEND = 'memory'