+++ molo/surveys/models.py
@@ -19,20 +19,41 @@ from molo.core.models import (
 )
 from molo.core.utils import generate_slug
 
-from wagtail.wagtailcore.models import Page
+from wagtail.wagtailcore.models import Page, Orderable
 from wagtail.wagtailadmin.edit_handlers import FieldPanel, InlinePanel, \
-    MultiFieldPanel
-from wagtailsurveys import models as surveys_models
+    MultiFieldPanel, StreamFieldPanel, PageChooserPanel, FieldRowPanel
+from wagtail.wagtailcore.fields import StreamField
+from wagtail.wagtailcore import blocks
+from wagtail.wagtailimages.blocks import ImageChooserBlock
+from wagtail.wagtailimages.edit_handlers import ImageChooserPanel
 
+from molo.core.blocks import MarkDownBlock
+from wagtailsurveys import models as surveys_models
+from django.db.models import Q
+from django.http import Http404
+from django.utils.translation import ugettext_lazy as _
+
+from wagtail_personalisation.adapters import get_segment_adapter
+from wagtailsurveys.models import AbstractFormField
+from .rules import SurveySubmissionDataRule, GroupMembershipRule  # noqa
 
 # See docs: https://github.com/torchbox/wagtailsurveys
 SectionPage.subpage_types += ['surveys.MoloSurveyPage']
 ArticlePage.subpage_types += ['surveys.MoloSurveyPage']
+ArticlePage.parent_page_types += ['surveys.TermsAndConditionsIndexPage']
+
+
+class TermsAndConditionsIndexPage(
+        TranslatablePageMixinNotRoutable, Page, PreventDeleteMixin):
+    parent_page_types = ['surveys.SurveysIndexPage']
+    subpage_types = ['core.ArticlePage']
 
 
 class SurveysIndexPage(Page, PreventDeleteMixin):
     parent_page_types = ['core.Main']
-    subpage_types = ['surveys.MoloSurveyPage']
+    subpage_types = [
+        'surveys.MoloSurveyPage', 'surveys.PersonalisableSurvey',
+        'surveys.TermsAndConditionsIndexPage']
 
     def copy(self, *args, **kwargs):
         site = kwargs['to'].get_site()
@@ -42,7 +63,7 @@ class SurveysIndexPage(Page, PreventDeleteMixin):
 
 
 @receiver(index_pages_after_copy, sender=Main)
-def create_survey_index_page(sender, instance, **kwargs):
+def create_survey_index_pages(sender, instance, **kwargs):
     if not instance.get_children().filter(
             title='Surveys').exists():
         survey_index = SurveysIndexPage(
@@ -59,8 +80,24 @@ class MoloSurveyPage(
     subpage_types = []
 
     intro = TextField(blank=True)
+    image = models.ForeignKey(
+        'wagtailimages.Image',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+'
+    )
+    content = StreamField([
+        ('heading', blocks.CharBlock(classname="full title")),
+        ('paragraph', MarkDownBlock()),
+        ('image', ImageChooserBlock()),
+        ('list', blocks.ListBlock(blocks.CharBlock(label="Item"))),
+        ('numbered_list', blocks.ListBlock(blocks.CharBlock(label="Item"))),
+        ('page', blocks.PageChooserBlock()),
+    ], null=True, blank=True)
     thank_you_text = TextField(blank=True)
-
+    submit_text = TextField(blank=True)
+    homepage_button_text = TextField(blank=True)
     allow_anonymous_submissions = BooleanField(
         default=False,
         help_text='Check this to allow users who are NOT logged in to complete'
@@ -77,6 +114,12 @@ class MoloSurveyPage(
         help_text='Whether to show the survey results to the user after they'
                   ' have submitted their answer(s).'
     )
+    show_results_as_percentage = BooleanField(
+        default=False,
+        help_text='Whether to show the survey results to the user after they'
+                  ' have submitted their answer(s) as a percentage or as'
+                  ' a number.'
+    )
 
     multi_step = BooleanField(
         default=False,
@@ -85,10 +128,34 @@ class MoloSurveyPage(
                   ' a time, instead of all at once.'
     )
 
+    display_survey_directly = BooleanField(
+        default=False,
+        verbose_name='Display Question Directly',
+        help_text='This is similar to polls, in which the questions are '
+                  'displayed directly on the page, instead of displaying '
+                  'a link to another page to complete the survey.'
+
+    )
+    your_words_competition = BooleanField(
+        default=False,
+        verbose_name='Is YourWords Competition',
+        help_text='This will display the correct template for yourwords'
+    )
+    extra_style_hints = models.TextField(
+        default='',
+        null=True, blank=True,
+        help_text=_(
+            "Styling options that can be applied to this page "
+            "and all its descendants"))
     content_panels = surveys_models.AbstractSurvey.content_panels + [
         FieldPanel('intro', classname='full'),
+        ImageChooserPanel('image'),
+        StreamFieldPanel('content'),
         InlinePanel('survey_form_fields', label='Form fields'),
         FieldPanel('thank_you_text', classname='full'),
+        FieldPanel('submit_text', classname='full'),
+        FieldPanel('homepage_button_text', classname='full'),
+        InlinePanel('terms_and_conditions', label="Terms and Conditions"),
     ]
 
     settings_panels = surveys_models.AbstractSurvey.settings_panels + [
@@ -96,9 +163,22 @@ class MoloSurveyPage(
             FieldPanel('allow_anonymous_submissions'),
             FieldPanel('allow_multiple_submissions_per_user'),
             FieldPanel('show_results'),
-            FieldPanel('multi_step')
-        ], heading='Survey Settings')
+            FieldPanel('show_results_as_percentage'),
+            FieldPanel('multi_step'),
+            FieldPanel('display_survey_directly'),
+            FieldPanel('your_words_competition'),
+        ], heading='Survey Settings'),
+        MultiFieldPanel(
+            [FieldRowPanel(
+                [FieldPanel('extra_style_hints')], classname="label-above")],
+            "Meta")
     ]
+
+    def get_effective_extra_style_hints(self):
+        return self.extra_style_hints
+
+    def get_effective_image(self):
+        return self.image
 
     def get_data_fields(self):
         data_fields = [
@@ -251,6 +331,20 @@ class MoloSurveyPage(
         return super(MoloSurveyPage, self).serve(request, *args, **kwargs)
 
 
+class SurveyArticlePage(Orderable):
+    page = ParentalKey(MoloSurveyPage, related_name='terms_and_conditions')
+    terms_and_conditions = models.ForeignKey(
+        'wagtailcore.Page',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+        help_text=_('Terms and Conditions')
+    )
+    panels = [PageChooserPanel(
+        'terms_and_conditions', 'core.ArticlePage')]
+
+
 class MoloSurveyFormField(surveys_models.AbstractFormField):
     page = ParentalKey(MoloSurveyPage, related_name='survey_form_fields')
 
@@ -259,6 +353,14 @@ class MoloSurveySubmission(surveys_models.AbstractFormSubmission):
     user = models.ForeignKey(
         settings.AUTH_USER_MODEL, on_delete=models.CASCADE, null=True
     )
+    article_page = models.ForeignKey(
+        'core.ArticlePage',
+        null=True,
+        blank=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+        help_text='Page to which the entry was converted to'
+    )
 
     def get_data(self):
         form_data = super(MoloSurveySubmission, self).get_data()
@@ -266,3 +368,114 @@ class MoloSurveySubmission(surveys_models.AbstractFormSubmission):
             'username': self.user.username if self.user else 'Anonymous',
         })
         return form_data
+
+
+# Personalised Surveys
+def get_personalisable_survey_content_panels():
+    """
+    Replace panel for "survey_form_fields" with
+    panel pointing to the custom form field model.
+    """
+    panels = [
+        FieldPanel('segment')
+    ]
+
+    for panel in MoloSurveyPage.content_panels:
+        if isinstance(panel, InlinePanel) and \
+                panel.relation_name == 'survey_form_fields':
+            panel = InlinePanel('personalisable_survey_form_fields',
+                                label=_('personalisable form fields'))
+        panels.append(panel)
+
+    return panels
+
+
+class PersonalisableSurvey(MoloSurveyPage):
+    """
+    Survey page that enables form fields to be segmented with
+    wagtail-personalisation.
+    """
+    segment = models.ForeignKey('wagtail_personalisation.Segment',
+                                on_delete=models.SET_NULL, blank=True,
+                                null=True,
+                                help_text=_(
+                                    'Leave it empty to show this survey'
+                                    'to every user.'))
+    content_panels = get_personalisable_survey_content_panels()
+    template = MoloSurveyPage.template
+
+    class Meta:
+        verbose_name = _('personalisable survey')
+
+    def get_form_fields(self):
+        """Get form fields for particular segments."""
+        # Get only segmented form fields if serve() has been called
+        # (because the page is being seen by user on the front-end)
+        if hasattr(self, 'request'):
+            user_segments_ids = [s.id for s in get_segment_adapter(
+                self.request).get_segments()]
+
+            return self.personalisable_survey_form_fields.filter(
+                Q(segment=None) | Q(segment_id__in=user_segments_ids)
+            )
+
+        # Return all form fields if there's no request passed
+        # (used on the admin site so serve() will not be called).
+        return self.personalisable_survey_form_fields \
+                   .select_related('segment')
+
+    def get_data_fields(self):
+        """
+        Get survey's form field's labels with segment names
+        if there's one associated.
+        """
+        data_fields = [
+            ('created_at', _('Submission Date')),
+        ]
+
+        # Add segment name to a field label if it is segmented.
+        for field in self.get_form_fields():
+            label = field.label
+
+            if field.segment:
+                label = '%s (%s)' % (label, field.segment.name)
+
+            data_fields.append((field.clean_name, label))
+
+        return data_fields
+
+    def serve(self, request, *args, **kwargs):
+        # We need request data in self.get_form_fields() to perform
+        # segmentation.
+        # TODO(tmkn): This is quite hacky, need to come up with better solution
+        self.request = request
+
+        # Check whether it is segmented and raise 404 if segments do not match
+        if self.segment_id and get_segment_adapter(request).get_segment_by_id(
+                self.segment_id) is None:
+            raise Http404("Survey does not match your segments.")
+
+        return super(PersonalisableSurvey, self).serve(
+            request, *args, **kwargs)
+
+
+class PersonalisableSurveyFormField(AbstractFormField):
+    """
+    Form field that has a segment assigned.
+    """
+    page = ParentalKey(PersonalisableSurvey, on_delete=models.CASCADE,
+                       related_name='personalisable_survey_form_fields')
+    segment = models.ForeignKey(
+        'wagtail_personalisation.Segment',
+        on_delete=models.PROTECT, blank=True, null=True,
+        help_text=_('Leave it empty to show this field to every user.'))
+
+    panels = [
+        FieldPanel('segment')
+    ] + AbstractFormField.panels
+
+    def __str__(self):
+        return '%s - %s' % (self.page, self.label)
+
+    class Meta:
+        verbose_name = _('personalisable form field')