+++ molo/surveys/wagtail_hooks.py
@@ -1,6 +1,19 @@
+from django.conf import settings
+from django.utils.html import format_html_join
+from django.contrib.auth.models import User
 
 from wagtail.wagtailcore import hooks
-from django.contrib.auth.models import User
+
+from django.conf.urls import url
+from django.core.urlresolvers import reverse
+from django.utils.translation import ugettext as _
+
+from wagtail.wagtailadmin.menu import MenuItem
+
+from molo.surveys.models import MoloSurveyPage, SurveyArticlePage
+from molo.core.models import ArticlePage
+
+from .views import create
 
 
 @hooks.register('construct_main_menu')
@@ -9,3 +22,53 @@ def show_surveys_entries_for_users_have_access(request, menu_items):
             pk=request.user.pk, groups__name='Moderators').exists():
         menu_items[:] = [
             item for item in menu_items if item.name != 'surveys']
+
+
+@hooks.register('insert_global_admin_js')
+def global_admin_js():
+    js_files = [
+        'js/survey-admin.js',
+    ]
+
+    js_includes = format_html_join(
+        '\n', '<script src="{0}{1}"></script>',
+        ((settings.STATIC_URL, filename) for filename in js_files)
+    )
+    return js_includes
+
+
+# CSV creation hooks
+
+
+@hooks.register('register_admin_urls')
+def register_csv_group_creation_urlconf():
+    return [
+        url('csv-group-creation/$', create, name='csv-group-creation'),
+    ]
+
+
+@hooks.register('register_settings_menu_item')
+def register_csv_group_creation_menu_item():
+    return MenuItem(
+        _('CSV group creation'),
+        reverse('csv-group-creation'),
+        classnames='icon icon-group', order=601)
+
+
+@hooks.register('after_copy_page')
+def create_new_page_relations(request, page, new_page):
+    if page and new_page:
+        if new_page.get_descendants().count() >= \
+                page.get_descendants().count():
+            for survey in MoloSurveyPage.objects.descendant_of(new_page):
+
+                # replace old terms and conditions with new one, if it exists
+                relations = SurveyArticlePage.objects.filter(page=survey)
+                for relation in relations:
+                    if relation.terms_and_conditions:
+                        new_article = ArticlePage.objects.descendant_of(
+                            new_page).filter(
+                                slug=relation.terms_and_conditions.slug)\
+                            .first()
+                        relation.terms_and_conditions = new_article
+                        relation.save()